reader-alt: query! macro
## Description

Introduce a new `query!` macro that works similarly to the existing
`sql!` macro, but is responsible for building whole queries, not just
snippets.

It improves on `BoxedSqlQuery` which is built into `diesel`, and
`RawQuery` which we used for similar reasons in OG GraphQL, in the
following ways:

- Unlike `BoxedSqlQuery` it supports interpolating one query into
  another.
- Unlike `RawQuery`, it supports binding parameters of all SQL types
  (not just `Text`), which avoids having to use `format!` to interpolate
  values in.
- Unlike both, `query!` supports appending one query onto another
  (correctly handling bind parameter ordering).

This abstraction is important for building the kinds of queries we need
for pagination (where cursors impose extra filters).

## Test plan

New unit tests for `Query` API:

```
sui$ cargo nextest run -p sui-pg-db
```

The `query!` macro will be used in a later change.