indexer-alt: retry all checkpoint fetch errors
## Description

Previously, the indexing framework would only retry a subset of remote
checkpoint fetching errors, and would wind down the indexer in response
to others.

This change makes it so that all errors (except "not found" errors) are
treated as transient and are retried with an exponential back-off. It
also simplifies the logic around cancellation, pulling it out from the
ingestion client, and introducing an outer, wrapping `select!` statement
that checks for cancellation instead.

## Test plan

Updtaed tests:

```
sui$ cargo nextest run         \
  -p sui-indexer-alt-framework \
  -p sui-indexer-alt           \
  -p sui-indexer-alt-e2e-tests
```